remove unnecessary file

+++ src/Launcher/Launcher.go
@@ -20,7 +20,9 @@ func main() {
 		}
 	}
 
-	fmt.Printf("good rate = %.3f, bad rate = %.3f\n", good/good+bad, bad/good+bad)
+	fmt.Printf("good rate = %.3f, bad rate = %.3f\n",
+		float64(good)/float64(good+bad),
+		float64(bad)/float64(good+bad))
 
 	fmt.Println("Game finished!")
 }
